+++ core/src/builtin_widgets/keep_alive.rs
@@ -1,29 +1,29 @@
 use crate::prelude::*;
 
-/// A widget that can delay drop its child until the `delay_drop_until` field be
-/// set to `true`.
+/// A widget will be painted event if it has been dispose until the `keep_alive`
+/// field be set to `false` or its parent is dropped.
 ///
 /// This widget not effect the widget lifecycle, if the widget is dispose but
-/// the `delay_drop_until` is `false`, it's not part of the widget tree anymore
+/// the `keep_alive` is `true`, it's not part of the widget tree anymore
 /// but not drop immediately, is disposed in `logic`, but not release resource.
 /// It's be isolated from the widget tree and can layout and paint normally.
 ///
-/// Once the `delay_drop_until` field be set to `true`, the widget will be
+/// Once the `keep_alive` field be set to `false`, the widget will be
 /// dropped.
 ///
 /// It's useful when you need run a leave animation for a widget.
 #[derive(Query, Default)]
-pub struct DelayDrop {
-  pub delay_drop_until: bool,
+pub struct KeepAlive {
+  pub keep_alive: bool,
 }
 
-impl Declare for DelayDrop {
+impl Declare for KeepAlive {
   type Builder = FatObj<()>;
   #[inline]
   fn declarer() -> Self::Builder { FatObj::new(()) }
 }
 
-impl ComposeChild for DelayDrop {
+impl ComposeChild for KeepAlive {
   type Child = Widget;
   #[inline]
   fn compose_child(this: impl StateWriter<Value = Self>, child: Self::Child) -> impl WidgetBuilder {
@@ -45,8 +45,8 @@ mod tests {
   fn smoke() {
     reset_test_env!();
 
-    let delay_drop = Stateful::new(false);
-    let c_delay_drop = delay_drop.clone_writer();
+    let keep_alive = Stateful::new(true);
+    let c_keep_alive = keep_alive.clone_writer();
     let remove_widget = Stateful::new(false);
     let c_remove_widget = remove_widget.clone_writer();
     let mut wnd = TestWindow::new(fn_widget! {
@@ -55,7 +55,7 @@ mod tests {
           Void.build(ctx!())
         } else {
           FatObj::new(Void)
-            .delay_drop_until(pipe!(*$delay_drop))
+            .keep_alive(pipe!(*$keep_alive))
             .build(ctx!())
         }
       }
@@ -73,7 +73,7 @@ mod tests {
     wnd.draw_frame();
     assert!(!root.is_dropped(&tree_arena(&wnd)));
 
-    *c_delay_drop.write() = true;
+    *c_keep_alive.write() = false;
     wnd.draw_frame();
     assert!(root.is_dropped(&tree_arena(&wnd)));
   }